:recycle:Refactor: refactor backtest transaction history code split 백테스트 매매내역 컴포넌트 분리
